add new dependency

+++ src/main/java/io/github/mmpodkanski/computershop/cart/CartFacade.java
@@ -5,8 +5,9 @@ import io.github.mmpodkanski.computershop.cart.dto.CartItemDto;
 import io.github.mmpodkanski.computershop.customer.Customer;
 import io.github.mmpodkanski.computershop.exception.ApiBadRequestException;
 import io.github.mmpodkanski.computershop.exception.ApiNotFoundException;
+import io.github.mmpodkanski.computershop.product.Product;
+import io.github.mmpodkanski.computershop.product.ProductFactory;
 import io.github.mmpodkanski.computershop.product.ProductQueryRepository;
-import io.github.mmpodkanski.computershop.product.dto.ProductDto;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -16,61 +17,66 @@ public class CartFacade {
     private final CartItemQueryRepository queryRepository;
     private final ProductQueryRepository productQueryRepository;
     private final CartFactory cartFactory;
+    private final ProductFactory productFactory;
 
     CartFacade(
             final CartItemRepository repository,
             final CartItemQueryRepository queryRepository,
             final ProductQueryRepository productQueryRepository,
-            final CartFactory cartFactory
+            final CartFactory cartFactory,
+            final ProductFactory productFactory
     ) {
         this.repository = repository;
         this.queryRepository = queryRepository;
         this.productQueryRepository = productQueryRepository;
         this.cartFactory = cartFactory;
+        this.productFactory = productFactory;
     }
 
-    public CartItemDto addToCart(AddToCartDto addToCartDto, Customer customer) {
-        var productDto = getProductFromCart(addToCartDto);
+    CartItemDto addToCart(AddToCartDto addToCartDto, Customer customer) {
+        var product = productQueryRepository.findById(addToCartDto.getProductId())
+                .orElseThrow(() -> new ApiNotFoundException("Product with that id not exists: " + addToCartDto.getProductId()));
 
-        var itemCart = cartFactory.toEntity(addToCartDto, customer, productDto);
-        return cartFactory.toDto(repository.save(itemCart));
-    }
+        var cartItem = queryRepository.findByProductAndCustomer(product, customer)
+                .map(cart -> {
+                    int finalQuantity = addToCartDto.getQuantity() + cart.getQuantity();
+                    checkIfProductIsAvailable(product, finalQuantity);
+                    cart.setQuantity(finalQuantity);
+                    return cart;
+                })
+                .orElseGet(() -> {
+                    checkIfProductIsAvailable(product, addToCartDto.getQuantity());
+                    return cartFactory.toEntity(addToCartDto, customer, productFactory.toDto(product));
+                });
 
-    // TODO: change repository method signature (should be only save methods with one exception)
-    @Transactional
-    public void updateCartItem(AddToCartDto addToCartDto, int cartId, Customer customer) {
-        var productDto = getProductFromCart(addToCartDto);
-        var cart = queryRepository.findByIdAndCustomer(cartId, customer);
-
-        if (!productDto.getCode().equals(cart.getProduct().getCode())) {
-            throw new ApiBadRequestException("Product is other than before!");
-        }
-        cart.setQuantity(addToCartDto.getQuantity());
+        return cartFactory.toDto(repository.save(cartItem));
     }
 
     @Transactional
     public void deleteCartItem(int id, Customer customer) {
-        if (!queryRepository.existsById(id))
-            throw new ApiNotFoundException("Cart with that id not exists : " + id);
+        if (!queryRepository.existsById(id)) {
+            throw new ApiNotFoundException("Cart with that id not exists: " + id);
+        }
 
-        repository.deleteByIdAndCustomer(id, customer);
+        queryRepository.findByIdAndCustomer(id, customer)
+                .ifPresent(cart -> {
+                    if(cart.getQuantity() > 1) {
+                        cart.decreaseQuantity();
+                    } else {
+                        repository.deleteByIdAndCustomer(id, customer);
+                    }
+                });
     }
 
     @Transactional
-    public void deleteCartItems(Customer customer) {
+    public void deleteAllCartItems(Customer customer) {
         repository.deleteAllByCustomer(customer);
     }
 
-
-    private ProductDto getProductFromCart(AddToCartDto addToCartDto) {
-        var productId = addToCartDto.getProductId();
-        var productDto = productQueryRepository.findDtoById(productId)
-                .orElseThrow(() -> new ApiNotFoundException("Product with that id not exists: " + productId));
-
-        if (productDto.getQuantity() < addToCartDto.getQuantity()) {
+    private void checkIfProductIsAvailable(Product product, int quantity) {
+        if (product.getQuantity() < quantity) {
             throw new ApiBadRequestException("You cant add more product than is available!");
         }
-        return productDto;
     }
 
 }
